Remove old traits for pairings. Create new pairings folder.
