+++ math/src/elliptic_curve/short_weierstrass/mod.rs
@@ -1,7 +1,7 @@
 /// Implementation of particular cases of elliptic curves.
 pub mod curves;
 /// Pairing algorithms
-pub mod pairing;
+pub mod pairings;
 /// Structs for points
 pub mod point;
 /// Common behaviour for Elliptic curves.
